Extract common code in marker displays into marker_common
