+++ rviz_default_plugins/include/rviz_default_plugins/displays/marker/markers/arrow_marker.hpp
@@ -59,7 +59,7 @@ class ArrowMarker : public MarkerBase
 public:
   RVIZ_DEFAULT_PLUGINS_PUBLIC
   ArrowMarker(
-    MarkerDisplay * owner, rviz_common::DisplayContext * context, Ogre::SceneNode * parent_node);
+    MarkerCommon * owner, rviz_common::DisplayContext * context, Ogre::SceneNode * parent_node);
 
   RVIZ_DEFAULT_PLUGINS_PUBLIC
   ~ArrowMarker() override = default;
