+++ rviz_default_plugins/include/rviz_default_plugins/displays/marker/markers/line_list_marker.hpp
@@ -49,7 +49,7 @@ class LineListMarker : public LineMarkerBase
 public:
   RVIZ_DEFAULT_PLUGINS_PUBLIC
   LineListMarker(
-    MarkerDisplay * owner, rviz_common::DisplayContext * context, Ogre::SceneNode * parent_node);
+    MarkerCommon * owner, rviz_common::DisplayContext * context, Ogre::SceneNode * parent_node);
   RVIZ_DEFAULT_PLUGINS_PUBLIC
   ~LineListMarker() override = default;
