+++ rviz_default_plugins/include/rviz_default_plugins/displays/marker/markers/mesh_resource_marker.hpp
@@ -62,7 +62,7 @@ class MeshResourceMarker : public MarkerBase
 public:
   RVIZ_DEFAULT_PLUGINS_PUBLIC
   MeshResourceMarker(
-    MarkerDisplay * owner, rviz_common::DisplayContext * context, Ogre::SceneNode * parent_node);
+    MarkerCommon * owner, rviz_common::DisplayContext * context, Ogre::SceneNode * parent_node);
 
   RVIZ_DEFAULT_PLUGINS_PUBLIC
   ~MeshResourceMarker() override;
