+++ rviz_default_plugins/include/rviz_default_plugins/displays/marker/markers/points_marker.hpp
@@ -53,7 +53,7 @@ class PointsMarker : public MarkerBase
 public:
   RVIZ_DEFAULT_PLUGINS_PUBLIC
   PointsMarker(
-    MarkerDisplay * owner, rviz_common::DisplayContext * context, Ogre::SceneNode * parent_node);
+    MarkerCommon * owner, rviz_common::DisplayContext * context, Ogre::SceneNode * parent_node);
 
   RVIZ_DEFAULT_PLUGINS_PUBLIC
   ~PointsMarker() override;
