+++ rviz_default_plugins/include/rviz_default_plugins/displays/marker/markers/shape_marker.hpp
@@ -53,7 +53,7 @@ class ShapeMarker : public MarkerBase
 public:
   RVIZ_DEFAULT_PLUGINS_PUBLIC
   ShapeMarker(
-    MarkerDisplay * owner, rviz_common::DisplayContext * context, Ogre::SceneNode * parent_node);
+    MarkerCommon * owner, rviz_common::DisplayContext * context, Ogre::SceneNode * parent_node);
 
   RVIZ_DEFAULT_PLUGINS_PUBLIC
   S_MaterialPtr getMaterials() override;
