+++ rviz_default_plugins/include/rviz_default_plugins/displays/marker/markers/text_view_facing_marker.hpp
@@ -56,7 +56,7 @@ class TextViewFacingMarker : public MarkerBase
 public:
   RVIZ_DEFAULT_PLUGINS_PUBLIC
   TextViewFacingMarker(
-    MarkerDisplay * owner, rviz_common::DisplayContext * context, Ogre::SceneNode * parent_node);
+    MarkerCommon * owner, rviz_common::DisplayContext * context, Ogre::SceneNode * parent_node);
 
   RVIZ_DEFAULT_PLUGINS_PUBLIC
   ~TextViewFacingMarker() override;
