+++ rviz_default_plugins/include/rviz_default_plugins/displays/marker/markers/triangle_list_marker.hpp
@@ -72,7 +72,7 @@ class TriangleListMarker : public MarkerBase
 public:
   RVIZ_DEFAULT_PLUGINS_PUBLIC
   TriangleListMarker(
-    MarkerDisplay * owner, rviz_common::DisplayContext * context, Ogre::SceneNode * parent_node);
+    MarkerCommon * owner, rviz_common::DisplayContext * context, Ogre::SceneNode * parent_node);
 
   RVIZ_DEFAULT_PLUGINS_PUBLIC
   ~TriangleListMarker() override;
