+++ rviz_default_plugins/src/rviz_default_plugins/displays/marker/markers/line_marker_base.cpp
@@ -38,12 +38,11 @@
 #include <OgreQuaternion.h>
 #include <OgreSceneNode.h>
 
-#include "rviz_default_plugins/displays/marker/marker_display.hpp"
-#include "rviz_default_plugins/displays/marker/markers/marker_selection_handler.hpp"
-
-#include "rviz_common/display_context.hpp"
-#include "rviz_common/properties/status_property.hpp"
 #include "rviz_rendering/objects/billboard_line.hpp"
+#include "rviz_common/display_context.hpp"
+
+#include "rviz_default_plugins/displays/marker/marker_common.hpp"
+#include "rviz_default_plugins/displays/marker/markers/marker_selection_handler.hpp"
 
 namespace rviz_default_plugins
 {
@@ -53,7 +52,7 @@ namespace markers
 {
 
 LineMarkerBase::LineMarkerBase(
-  MarkerDisplay * owner, rviz_common::DisplayContext * context, Ogre::SceneNode * parent_node)
+  MarkerCommon * owner, rviz_common::DisplayContext * context, Ogre::SceneNode * parent_node)
 : MarkerBase(owner, context, parent_node), lines_(nullptr), has_per_point_color_(false)
 {}
