+++ rviz_default_plugins/src/rviz_default_plugins/displays/marker/markers/line_strip_marker.cpp
@@ -36,10 +36,10 @@
 #include <OgreQuaternion.h>
 #include <OgreSceneNode.h>
 
-#include "rviz_default_plugins/displays/marker/marker_display.hpp"
-
-#include "rviz_common/display_context.hpp"
 #include "rviz_rendering/objects/billboard_line.hpp"
+#include "rviz_common/display_context.hpp"
+
+#include "rviz_default_plugins/displays/marker/marker_common.hpp"
 
 namespace rviz_default_plugins
 {
@@ -49,7 +49,7 @@ namespace markers
 {
 
 LineStripMarker::LineStripMarker(
-  MarkerDisplay * owner, rviz_common::DisplayContext * context, Ogre::SceneNode * parent_node)
+  MarkerCommon * owner, rviz_common::DisplayContext * context, Ogre::SceneNode * parent_node)
 : LineMarkerBase(owner, context, parent_node)
 {}
