+++ rviz_default_plugins/src/rviz_default_plugins/displays/marker/markers/mesh_resource_marker.cpp
@@ -50,9 +50,11 @@
 #include <OgreTechnique.h>
 #include <OgreTextureManager.h>
 
-#include "rviz_default_plugins/displays/marker/marker_display.hpp"
-#include "rviz_default_plugins/displays/marker/markers/marker_selection_handler.hpp"
 #include "rviz_rendering/mesh_loader.hpp"
+#include "rviz_common/display_context.hpp"
+
+#include "rviz_default_plugins/displays/marker/marker_common.hpp"
+#include "rviz_default_plugins/displays/marker/markers/marker_selection_handler.hpp"
 
 namespace rviz_default_plugins
 {
@@ -62,7 +64,7 @@ namespace markers
 {
 
 MeshResourceMarker::MeshResourceMarker(
-  MarkerDisplay * owner, rviz_common::DisplayContext * context, Ogre::SceneNode * parent_node)
+  MarkerCommon * owner, rviz_common::DisplayContext * context, Ogre::SceneNode * parent_node)
 : MarkerBase(owner, context, parent_node), entity_(nullptr)
 {}
