+++ rviz_default_plugins/src/rviz_default_plugins/displays/marker/markers/points_marker.cpp
@@ -34,7 +34,7 @@
 
 #include "rviz_common/interaction/selection_manager.hpp"
 
-#include "rviz_default_plugins/displays/marker/marker_display.hpp"
+#include "rviz_default_plugins/displays/marker/marker_common.hpp"
 #include "rviz_default_plugins/displays/marker/markers/marker_selection_handler.hpp"
 
 namespace rviz_default_plugins
@@ -44,7 +44,7 @@ namespace displays
 namespace markers
 {
 PointsMarker::PointsMarker(
-  MarkerDisplay * owner, rviz_common::DisplayContext * context, Ogre::SceneNode * parent_node)
+  MarkerCommon * owner, rviz_common::DisplayContext * context, Ogre::SceneNode * parent_node)
 : MarkerBase(owner, context, parent_node),
   points_(nullptr)
 {}
