+++ rviz_default_plugins/src/rviz_default_plugins/displays/marker/markers/shape_marker.cpp
@@ -32,11 +32,11 @@
 
 #include <memory>
 
-#include "rviz_default_plugins/displays/marker/markers/marker_selection_handler.hpp"
-#include "rviz_default_plugins/displays/marker/marker_display.hpp"
-#include "rviz_common/display_context.hpp"
-#include "rviz_common/properties/status_property.hpp"
 #include "rviz_rendering/objects/shape.hpp"
+#include "rviz_common/display_context.hpp"
+
+#include "rviz_default_plugins/displays/marker/marker_common.hpp"
+#include "rviz_default_plugins/displays/marker/markers/marker_selection_handler.hpp"
 
 namespace rviz_default_plugins
 {
@@ -46,7 +46,7 @@ namespace markers
 {
 
 ShapeMarker::ShapeMarker(
-  MarkerDisplay * owner, rviz_common::DisplayContext * context, Ogre::SceneNode * parent_node)
+  MarkerCommon * owner, rviz_common::DisplayContext * context, Ogre::SceneNode * parent_node)
 : MarkerBase(owner, context, parent_node),
   shape_(nullptr)
 {}
