+++ rviz_default_plugins/test/rviz_default_plugins/displays/marker/marker_common_test.cpp
@@ -40,8 +40,9 @@
 
 #include "visualization_msgs/msg/marker.hpp"
 
-#include "rviz_default_plugins/displays/marker/marker_display.hpp"
-#include "rviz_default_plugins/displays/marker/markers/marker_factory.hpp"
+#include "rviz_common/display.hpp"
+
+#include "rviz_default_plugins/displays/marker/marker_common.hpp"
 #include "rviz_default_plugins/displays/marker/markers/arrow_marker.hpp"
 #include "rviz_default_plugins/displays/marker/markers/shape_marker.hpp"
 #include "rviz_default_plugins/displays/marker/markers/line_strip_marker.hpp"
@@ -58,21 +59,25 @@
 
 using namespace ::testing;  // NOLINT
 
-class MarkerDisplayFixture : public DisplayTestFixture
+class MarkerCommonFixture : public DisplayTestFixture
 {
 public:
-  MarkerDisplayFixture()
+  MarkerCommonFixture()
   {
-    factory_ = std::make_unique<rviz_default_plugins::displays::markers::MarkerFactory>();
-    factory_->initialize(
-      nullptr, context_.get(), scene_manager_->getRootSceneNode()->createChildSceneNode());
+    display_ = std::make_unique<rviz_common::Display>();
 
-    display_ = std::make_unique<rviz_default_plugins::displays::MarkerDisplay>(
-      std::move(factory_), context_.get());
+    auto scene_node = scene_manager_->getRootSceneNode()->createChildSceneNode();
+    common_ = std::make_unique<rviz_default_plugins::displays::MarkerCommon>(display_.get());
+    common_->initialize(context_.get(), scene_node);
   }
 
-  std::unique_ptr<rviz_default_plugins::displays::markers::MarkerFactory> factory_;
-  std::unique_ptr<rviz_default_plugins::displays::MarkerDisplay> display_;
+  ~MarkerCommonFixture() override
+  {
+    common_.reset();
+  }
+
+  std::unique_ptr<rviz_default_plugins::displays::MarkerCommon> common_;
+  std::unique_ptr<rviz_common::Display> display_;
 };
 
 visualization_msgs::msg::Marker::SharedPtr createSharedPtrMessage(
@@ -84,10 +89,10 @@ visualization_msgs::msg::Marker::SharedPtr createSharedPtrMessage(
   return std::make_shared<visualization_msgs::msg::Marker>(marker);
 }
 
-TEST_F(MarkerDisplayFixture, processMessage_creates_correct_marker_on_add_type) {
+TEST_F(MarkerCommonFixture, processMessage_creates_correct_marker_on_add_type) {
   mockValidTransform();
 
-  display_->processMessage(createSharedPtrMessage(
+  common_->processMessage(createSharedPtrMessage(
       visualization_msgs::msg::Marker::ADD, visualization_msgs::msg::Marker::TEXT_VIEW_FACING));
 
   auto text = rviz_rendering::findOneMovableText(scene_manager_->getRootSceneNode());
@@ -96,20 +101,20 @@ TEST_F(MarkerDisplayFixture, processMessage_creates_correct_marker_on_add_type) 
 }
 
 
-TEST_F(MarkerDisplayFixture, processMessage_deletes_correct_marker_on_delete_type) {
+TEST_F(MarkerCommonFixture, processMessage_deletes_correct_marker_on_delete_type) {
   mockValidTransform();
 
-  display_->processMessage(createSharedPtrMessage(
+  common_->processMessage(createSharedPtrMessage(
       visualization_msgs::msg::Marker::ADD,
       visualization_msgs::msg::Marker::TEXT_VIEW_FACING,
       0));
 
-  display_->processMessage(createSharedPtrMessage(
+  common_->processMessage(createSharedPtrMessage(
       visualization_msgs::msg::Marker::ADD,
       visualization_msgs::msg::Marker::POINTS,
       1));
 
-  display_->processMessage(createSharedPtrMessage(
+  common_->processMessage(createSharedPtrMessage(
       visualization_msgs::msg::Marker::DELETE,
       visualization_msgs::msg::Marker::TEXT_VIEW_FACING,
       0));
@@ -118,20 +123,20 @@ TEST_F(MarkerDisplayFixture, processMessage_deletes_correct_marker_on_delete_typ
   ASSERT_TRUE(rviz_rendering::findOnePointCloud(scene_manager_->getRootSceneNode()));
 }
 
-TEST_F(MarkerDisplayFixture, processMessage_with_deleteall_deletes_all_markers) {
+TEST_F(MarkerCommonFixture, processMessage_with_deleteall_deletes_all_markers) {
   mockValidTransform();
 
-  display_->processMessage(createSharedPtrMessage(
+  common_->processMessage(createSharedPtrMessage(
       visualization_msgs::msg::Marker::ADD,
       visualization_msgs::msg::Marker::TEXT_VIEW_FACING,
       0));
 
-  display_->processMessage(createSharedPtrMessage(
+  common_->processMessage(createSharedPtrMessage(
       visualization_msgs::msg::Marker::ADD,
       visualization_msgs::msg::Marker::POINTS,
       1));
 
-  display_->processMessage(createSharedPtrMessage(
+  common_->processMessage(createSharedPtrMessage(
       visualization_msgs::msg::Marker::DELETEALL,
       visualization_msgs::msg::Marker::TEXT_VIEW_FACING,
       3));
@@ -140,47 +145,47 @@ TEST_F(MarkerDisplayFixture, processMessage_with_deleteall_deletes_all_markers) 
   ASSERT_FALSE(rviz_rendering::findOnePointCloud(scene_manager_->getRootSceneNode()));
 }
 
-TEST_F(MarkerDisplayFixture, proccesMessage_add_all_markers_correctly) {
+TEST_F(MarkerCommonFixture, proccesMessage_add_all_markers_correctly) {
   mockValidTransform();
 
   auto marker = createSharedPtrMessage(
     visualization_msgs::msg::Marker::ADD, visualization_msgs::msg::Marker::TEXT_VIEW_FACING);
-  display_->processMessage(marker);
+  common_->processMessage(marker);
   ASSERT_TRUE(rviz_rendering::findOneMovableText(scene_manager_->getRootSceneNode()));
 
-  display_->deleteAllMarkers();
+  common_->deleteAllMarkers();
 
   marker->type = visualization_msgs::msg::Marker::MESH_RESOURCE;
-  display_->processMessage(marker);
+  common_->processMessage(marker);
   ASSERT_TRUE(rviz_rendering::findEntityByMeshName(
       scene_manager_->getRootSceneNode(), marker->mesh_resource));
 
-  display_->deleteAllMarkers();
+  common_->deleteAllMarkers();
 
   marker->type = visualization_msgs::msg::Marker::ARROW;
-  display_->processMessage(marker);
+  common_->processMessage(marker);
   ASSERT_TRUE(rviz_rendering::findOneArrow(scene_manager_->getRootSceneNode()));
 
-  display_->deleteAllMarkers();
+  common_->deleteAllMarkers();
 
   marker->type = visualization_msgs::msg::Marker::LINE_LIST;
-  display_->processMessage(marker);
+  common_->processMessage(marker);
   ASSERT_TRUE(rviz_rendering::findOneBillboardChain(scene_manager_->getRootSceneNode()));
 
-  display_->deleteAllMarkers();
+  common_->deleteAllMarkers();
 
   marker->type = visualization_msgs::msg::Marker::LINE_STRIP;
-  display_->processMessage(marker);
+  common_->processMessage(marker);
   ASSERT_TRUE(rviz_rendering::findOneBillboardChain(scene_manager_->getRootSceneNode()));
 
-  display_->deleteAllMarkers();
+  common_->deleteAllMarkers();
 
   marker->type = visualization_msgs::msg::Marker::CYLINDER;
-  display_->processMessage(marker);
+  common_->processMessage(marker);
   ASSERT_TRUE(rviz_rendering::findEntityByMeshName(
       scene_manager_->getRootSceneNode(), "rviz_cylinder.mesh"));
 
-  display_->deleteAllMarkers();
+  common_->deleteAllMarkers();
 
   geometry_msgs::msg::Point point;
   point.x = 0;
@@ -188,23 +193,23 @@ TEST_F(MarkerDisplayFixture, proccesMessage_add_all_markers_correctly) {
   point.z = 0;
   marker->type = visualization_msgs::msg::Marker::TRIANGLE_LIST;
   marker->points.push_back(point);
-  display_->processMessage(marker);
+  common_->processMessage(marker);
   ASSERT_TRUE(rviz_rendering::findOneManualObject(scene_manager_->getRootSceneNode()));
 
-  display_->deleteAllMarkers();
+  common_->deleteAllMarkers();
 }
 
-TEST_F(MarkerDisplayFixture, processMessage_adds_two_markers_of_same_type_if_ids_are_different) {
+TEST_F(MarkerCommonFixture, processMessage_adds_two_markers_of_same_type_if_ids_are_different) {
   mockValidTransform();
 
-  display_->processMessage(createSharedPtrMessage(
+  common_->processMessage(createSharedPtrMessage(
       visualization_msgs::msg::Marker::ADD,
       visualization_msgs::msg::Marker::ARROW,
       0));
 
   EXPECT_THAT(rviz_rendering::findAllArrows(scene_manager_->getRootSceneNode()), SizeIs(1));
 
-  display_->processMessage(createSharedPtrMessage(
+  common_->processMessage(createSharedPtrMessage(
       visualization_msgs::msg::Marker::ADD,
       visualization_msgs::msg::Marker::ARROW,
       1));
@@ -212,34 +217,34 @@ TEST_F(MarkerDisplayFixture, processMessage_adds_two_markers_of_same_type_if_ids
   EXPECT_THAT(rviz_rendering::findAllArrows(scene_manager_->getRootSceneNode()), SizeIs(2));
 }
 
-TEST_F(MarkerDisplayFixture,
+TEST_F(MarkerCommonFixture,
   processMessage_adds_two_markers_of_same_type_if_namespaces_are_different) {
   mockValidTransform();
 
   auto marker = createSharedPtrMessage(
     visualization_msgs::msg::Marker::ADD, visualization_msgs::msg::Marker::ARROW);
 
-  display_->processMessage(marker);
+  common_->processMessage(marker);
 
   EXPECT_THAT(rviz_rendering::findAllArrows(scene_manager_->getRootSceneNode()), SizeIs(1));
 
   marker->ns = "new_ns";
-  display_->processMessage(marker);
+  common_->processMessage(marker);
 
   EXPECT_THAT(rviz_rendering::findAllArrows(scene_manager_->getRootSceneNode()), SizeIs(2));
 }
 
-TEST_F(MarkerDisplayFixture, processMessage_does_not_create_new_marker_if_id_already_exists) {
+TEST_F(MarkerCommonFixture, processMessage_does_not_create_new_marker_if_id_already_exists) {
   mockValidTransform();
 
-  display_->processMessage(createSharedPtrMessage(
+  common_->processMessage(createSharedPtrMessage(
       visualization_msgs::msg::Marker::ADD,
       visualization_msgs::msg::Marker::ARROW,
       0));
 
   EXPECT_THAT(rviz_rendering::findAllArrows(scene_manager_->getRootSceneNode()), SizeIs(1));
 
-  display_->processMessage(createSharedPtrMessage(
+  common_->processMessage(createSharedPtrMessage(
       visualization_msgs::msg::Marker::ADD,
       visualization_msgs::msg::Marker::ARROW,
       0));
@@ -247,36 +252,36 @@ TEST_F(MarkerDisplayFixture, processMessage_does_not_create_new_marker_if_id_alr
   EXPECT_THAT(rviz_rendering::findAllArrows(scene_manager_->getRootSceneNode()), SizeIs(1));
 }
 
-TEST_F(MarkerDisplayFixture, processMessage_updates_modified_marker) {
+TEST_F(MarkerCommonFixture, processMessage_updates_modified_marker) {
   mockValidTransform();
 
   auto marker = createSharedPtrMessage(
     visualization_msgs::msg::Marker::ADD, visualization_msgs::msg::Marker::TEXT_VIEW_FACING);
-  display_->processMessage(marker);
+  common_->processMessage(marker);
 
   auto before_update = rviz_rendering::findOneMovableText(scene_manager_->getRootSceneNode());
   ASSERT_TRUE(before_update);
   EXPECT_THAT(before_update->getCaption(), StrEq(marker->text));
 
   marker->text = "New text";
-  display_->processMessage(marker);
+  common_->processMessage(marker);
 
   auto after_update = rviz_rendering::findOneMovableText(scene_manager_->getRootSceneNode());
   ASSERT_TRUE(after_update);
   EXPECT_THAT(after_update->getCaption(), StrEq("New text"));
 }
 
-TEST_F(MarkerDisplayFixture, processMessage_using_modify_works_like_add) {
+TEST_F(MarkerCommonFixture, processMessage_using_modify_works_like_add) {
   mockValidTransform();
 
-  display_->processMessage(createSharedPtrMessage(
+  common_->processMessage(createSharedPtrMessage(
       visualization_msgs::msg::Marker::ADD,
       visualization_msgs::msg::Marker::ARROW,
       0));
 
   EXPECT_THAT(rviz_rendering::findAllArrows(scene_manager_->getRootSceneNode()), SizeIs(1));
 
-  display_->processMessage(createSharedPtrMessage(
+  common_->processMessage(createSharedPtrMessage(
       visualization_msgs::msg::Marker::MODIFY,
       visualization_msgs::msg::Marker::ARROW,
       0));
@@ -284,7 +289,7 @@ TEST_F(MarkerDisplayFixture, processMessage_using_modify_works_like_add) {
   EXPECT_THAT(rviz_rendering::findAllArrows(scene_manager_->getRootSceneNode()), SizeIs(1));
 }
 
-TEST_F(MarkerDisplayFixture, update_retransforms_frame_locked_messages) {
+TEST_F(MarkerCommonFixture, update_retransforms_frame_locked_messages) {
   auto marker = createSharedPtrMessage(
     visualization_msgs::msg::Marker::ADD, visualization_msgs::msg::Marker::POINTS);
   marker->frame_locked = true;
@@ -306,7 +311,7 @@ TEST_F(MarkerDisplayFixture, update_retransforms_frame_locked_messages) {
       SetArgReferee<4>(next_orientation),
       Return(true)));
 
-  display_->processMessage(marker);
+  common_->processMessage(marker);
 
   auto pointCloud = rviz_rendering::findOnePointCloud(scene_manager_->getRootSceneNode());
   ASSERT_TRUE(pointCloud);
@@ -314,14 +319,14 @@ TEST_F(MarkerDisplayFixture, update_retransforms_frame_locked_messages) {
   EXPECT_THAT(pointCloud->getParentSceneNode()->getOrientation(),
     QuaternionEq(starting_orientation));
 
-  display_->update(0, 0);
+  common_->update(0, 0);
 
   EXPECT_THAT(pointCloud->getParentSceneNode()->getPosition(), Vector3Eq(next_position));
   EXPECT_THAT(pointCloud->getParentSceneNode()->getOrientation(),
     QuaternionEq(next_orientation));
 }
 
-TEST_F(MarkerDisplayFixture, update_does_not_retransform_normal_messages) {
+TEST_F(MarkerCommonFixture, update_does_not_retransform_normal_messages) {
   auto marker = createSharedPtrMessage(
     visualization_msgs::msg::Marker::ADD, visualization_msgs::msg::Marker::POINTS);
   marker->frame_locked = false;
@@ -337,7 +342,7 @@ TEST_F(MarkerDisplayFixture, update_does_not_retransform_normal_messages) {
       Return(true)
     ));  // Test will fail if this function is called repeatedly
 
-  display_->processMessage(marker);
+  common_->processMessage(marker);
 
   auto pointCloud = rviz_rendering::findOnePointCloud(scene_manager_->getRootSceneNode());
   ASSERT_TRUE(pointCloud);
@@ -345,81 +350,81 @@ TEST_F(MarkerDisplayFixture, update_does_not_retransform_normal_messages) {
   EXPECT_THAT(pointCloud->getParentSceneNode()->getOrientation(),
     QuaternionEq(starting_orientation));
 
-  display_->update(0, 0);
+  common_->update(0, 0);
 
   EXPECT_THAT(pointCloud->getParentSceneNode()->getPosition(), Vector3Eq(starting_position));
   EXPECT_THAT(pointCloud->getParentSceneNode()->getOrientation(),
     QuaternionEq(starting_orientation));
 }
 
-TEST_F(MarkerDisplayFixture, processMessage_adds_new_namespace_for_message) {
+TEST_F(MarkerCommonFixture, processMessage_adds_new_namespace_for_message) {
   mockValidTransform();
 
-  ASSERT_THAT(display_->childAt(3)->numChildren(), Eq(0));
+  ASSERT_THAT(display_->childAt(0)->numChildren(), Eq(0));
 
   auto marker = createSharedPtrMessage(
     visualization_msgs::msg::Marker::ADD, visualization_msgs::msg::Marker::POINTS);
 
-  display_->processMessage(marker);
+  common_->processMessage(marker);
 
-  ASSERT_THAT(display_->childAt(3)->childAt(0)->getName().toStdString(), StrEq("test_ns"));
+  ASSERT_THAT(display_->childAt(0)->childAt(0)->getName().toStdString(), StrEq("test_ns"));
 }
 
-TEST_F(MarkerDisplayFixture, processMessage_does_not_add_new_namespace_if_already_present) {
+TEST_F(MarkerCommonFixture, processMessage_does_not_add_new_namespace_if_already_present) {
   mockValidTransform();
 
-  display_->processMessage(createSharedPtrMessage(
+  common_->processMessage(createSharedPtrMessage(
       visualization_msgs::msg::Marker::ADD, visualization_msgs::msg::Marker::POINTS));
 
-  ASSERT_THAT(display_->childAt(3)->numChildren(), Eq(1));
-  ASSERT_THAT(display_->childAt(3)->childAt(0)->getName().toStdString(), StrEq("test_ns"));
+  ASSERT_THAT(display_->childAt(0)->numChildren(), Eq(1));
+  ASSERT_THAT(display_->childAt(0)->childAt(0)->getName().toStdString(), StrEq("test_ns"));
 
-  display_->processMessage(createSharedPtrMessage(
+  common_->processMessage(createSharedPtrMessage(
       visualization_msgs::msg::Marker::ADD, visualization_msgs::msg::Marker::TEXT_VIEW_FACING));
 
-  ASSERT_THAT(display_->childAt(3)->numChildren(), Eq(1));
-  ASSERT_THAT(display_->childAt(3)->childAt(0)->getName().toStdString(), StrEq("test_ns"));
+  ASSERT_THAT(display_->childAt(0)->numChildren(), Eq(1));
+  ASSERT_THAT(display_->childAt(0)->childAt(0)->getName().toStdString(), StrEq("test_ns"));
 }
 
-TEST_F(MarkerDisplayFixture, onEnableChanged_in_namespace_removes_all_markers_in_that_namespace) {
+TEST_F(MarkerCommonFixture, onEnableChanged_in_namespace_removes_all_markers_in_that_namespace) {
   mockValidTransform();
 
   auto marker = createSharedPtrMessage(
     visualization_msgs::msg::Marker::ADD, visualization_msgs::msg::Marker::POINTS);
 
-  display_->processMessage(marker);
+  common_->processMessage(marker);
   marker->type = visualization_msgs::msg::Marker::TEXT_VIEW_FACING;
   marker->ns = "new_ns";
-  display_->processMessage(marker);
+  common_->processMessage(marker);
 
   EXPECT_TRUE(rviz_rendering::findOnePointCloud(scene_manager_->getRootSceneNode()));
   EXPECT_TRUE(rviz_rendering::findOneMovableText(scene_manager_->getRootSceneNode()));
 
   auto namespace_property = dynamic_cast<rviz_default_plugins::displays::MarkerNamespace *>(
-    display_->childAt(3)->childAt(0));
+    display_->childAt(0)->childAt(0));
   namespace_property->setValue(false);
 
   EXPECT_FALSE(rviz_rendering::findOnePointCloud(scene_manager_->getRootSceneNode()));
   EXPECT_TRUE(rviz_rendering::findOneMovableText(scene_manager_->getRootSceneNode()));
 }
 
-TEST_F(MarkerDisplayFixture, processMessage_does_not_add_message_with_disabled_namespace) {
+TEST_F(MarkerCommonFixture, processMessage_does_not_add_message_with_disabled_namespace) {
   mockValidTransform();
 
   auto marker = createSharedPtrMessage(
     visualization_msgs::msg::Marker::ADD, visualization_msgs::msg::Marker::POINTS);
 
   // this is necessary to initialize namespace as we don't load a config
-  display_->processMessage(marker);
+  common_->processMessage(marker);
 
   EXPECT_TRUE(rviz_rendering::findOnePointCloud(scene_manager_->getRootSceneNode()));
 
   auto namespace_property = dynamic_cast<rviz_default_plugins::displays::MarkerNamespace *>(
-    display_->childAt(3)->childAt(0));
+    display_->childAt(0)->childAt(0));
   namespace_property->setValue(false);
 
   marker->type = visualization_msgs::msg::Marker::TEXT_VIEW_FACING;
-  display_->processMessage(marker);
+  common_->processMessage(marker);
 
   EXPECT_FALSE(rviz_rendering::findOnePointCloud(scene_manager_->getRootSceneNode()));
   EXPECT_FALSE(rviz_rendering::findOneMovableText(scene_manager_->getRootSceneNode()));
